This submission addresses Trac tickets 15, 21, 22, 24, 25, & 26.
This corrects several of the inconveniences of building CS-MAP on Linux 64 bit systems.  Also note, that the perennial prolem of _wcsicmp has been resolved by writing our own equivalent: CS_wcsicmp in CS_ansi.c

Also note that the compilation failure on 64 bit gcc was due to the fact that a "long" in 64 bit gcc (and also Sun, by the way) is a 64 bit object.  CS-MAP uses the "long32_t" and "ulong32_t" wherever it is expecting a 32 bit integer value.  This useage should/must be adhered to in order to avoid similar problems in the future.