Trac Tickets #51 & 52.  These two defects required changes to a common file (NameMapper.csv) so they end up in one submission.
For ticket 51, the correction consists of changing the Autodesk name for EPSG entries 3097 thru 3101.

For Ticket 52, a new datum, "Ocotepeque 1935 A" (named Ocotepeque35a) was added, and new variations of the CR-N and CR-S coordinate systems were added which reference the new Ocotepeque 1935 datum.  The new systems imply additions to the Catergory.asc file, the NameMapper file, and the data files used to generate the NameMapper file.

Test cases for the new datum were added to the Test.DAT file.

Note that the actual parameters for the new Ocotepeque datum were chosen from five candidates.  The chosen values were those that gave the best results when used on an end user test data set.  The results were not great, about 8 meters off.  All the other definition candidates produced discrepancies of 100 meters.  The 'a' appendix to the name was added assuming that the other definition candidates may need to be added at some point in the future.