Fix a CsMap build issue...
The postbuild step of the Compiler project runs the CS_Comp program to create the
dictionaries.  The Release|Win32 version of CS_Comp run on all machines, both 32-bit
and 64-bit.  But the Release|x64 version will only run on 64-bit machines.

I've updated the post-build step for the Release|x64 configuration to only run
the CS_Comp program if the PROCESSOR_ARCHITECTURE is AMD64, which corresponds to
64-bit machines.