Trac ticket #65.  Update HARN dictionary definitions to make use of the new features implemented by RFC 2.
The code submission for RFC #2 failed to include updated dictionaries in which 
the HARN/HPGN implementation takes advantage of the new features made available 
by the RFC.  This submission corrects this matter.  The end result is:

1> approximately 55 new datum definitions,
2> deprecation of a datum definition
3> approximately 55 new geodetic transformation definitions
4> approximately 500 new coordinate reference system definitions
5> deprecation of about 400 coordinate reference system definitions
6> categories updated appropriately for the definition modifications.

Note, these changes to the dictionary source files were done programmatically.  This submission includes the code for those programs.