Major fix for CS-MAP Trac Ticket #104.
This submission includes the code for console programs which edit the following dictionaries, and the modified dictionaries themselves:
1> category.asc
2> coordsys.asc
3> datums.asc
4> GeodeticTransfotmation.asc, and the
5> NameMapper.csv database file.

These data files were modified as appropriate to deprecate and replace several datums and the coordinate systems and geodetic transformations which reference them.  The list is given in a table coded into the SevenParamaterFixVersion2.cpp module.  Tweaks to some of the test modules were made to produce accurate testing of the results.

In the past, modifications to the NameMapper.csv file were accomplished by editing the component .csv files maintained in the Data subdirectory and then rebuilding the NameMapper.csv file from scratch.  As in the last twelve months several edits to the NameMapper.csv file were made manually, this technique could no longer be used.  This submission includes some minor changes to the Name Mapper code which enable the name mapper database to be manipulated programatically (i.e. updated).  This was the process used in this case, and what will need to be used from now on.  The ....KeyNameMap.csv files in the Data subdirectory are no longer useful and should be deleted at some point in the near future.

Tested using the Tortoise Diff and ConsoleTestCpp modules.