Optimize loading performance of name mapper Fix ticket 165
At the first use of the name mapper, the mapping information is loaded from the huge (1.4MB) dictionary file 'NameMapper.csv'. This is time consuming. Currently in a standard environment the loading takes about 500ms.
The current submission tries to optimize the loading of the name mapper that means lowering the time used to load. This is done by loading the complete file at once into the memory to omit too many file access. Then the parsing is done from the in-memory data and not directly on the file. This brings down the loading performance on a standard environment to 300ms.
Of course more performance optimization could be done. But this submission is now one step.

Changes reviewed by Andre.