Handle duplicate (code) geodetic transformation definitions
Add also duplicate geodetic transformation definitions, same code, to the geodetic transformation definition collection (gx) like it is possible for the other definitions (CRS, datum, ellipsoid, path). This allows a consumer to find out his redundant definitions and delete them over the API.

It is not expected to keep redundant definitions in the dictionaries. But by mistake redundant definitions could come in. E.g. the use case when a definition exists in the user defined dictionary and an identical system one comes with a system update.

Fix ticket #169