And finally, the vertical shift files for Alaska, 2007 and 2011.
This should, finally, complete the RFC-7 submission.  As described in the RFC, the size of the binary data files will probably be an issue for many folks.  Having a 1 minute grid file for Alaska does sound pretty silly.  Anyway, now that the RFC is complete and available for evaluation, this issue can be discussed intelligently.