Trac Ticket 135 -- Submission corrects a problem on big endian machines where on dictionary writes, a swapped method code is used to determine the swapping to be applied to the remainder of the geodetic transformation definition being written.
This was corrected by adding a parameter to the previous CS_gxswp function which is TRUE is the parent function requesting the swap is a write.  To be compatible with the new csIoUtil.cpp module, two additional functions, CS_gcswpRd and CS_gxswpWr were added which simply call the primary CS_gxswp function with the appropriate value for the new parameter.

This developer does not have access to a big endian machine and could not test this  change thoroughly.  It was tested in a little endian environment and it was verified that this change did not break anything in that environment.